basic grade calculationi is done i think

+++ database/migrations/2025_09_11_074039_create_grades_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('grades', function (Blueprint $table) {
             $table->id();
-            $table->integer('grade');
+            $table->float('grade');
             $table->foreignId('dosen_id')->references('id')->on('users')->onDelete('cascade');
             $table->timestamps();
         });
